requirements.txt
pandas 
numpy 
matplotlib 
statsmodels 
scipy 
python-dateutil